optimized vtables in windows part hopefully fixed crash #3666 onSecond and onMinute ticks supports 64bit times first try to fix Hub reconnect (THX adrian_007 for some ideas)

+++ client/AdcHub.cpp
@@ -652,7 +652,7 @@ void AdcHub::on(Failed, const string& aLine) throw() {
 	Client::on(Failed(), aLine);
 }
 
-void AdcHub::on(Second, uint32_t aTick) throw() {
+void AdcHub::on(Second, uint64_t aTick) throw() {
 	Client::on(Second(), aTick);
 	if(state == STATE_NORMAL && (aTick > (getLastActivity() + 120*1000)) ) {
 		send("\n", 1);
