+++ client/ConnectionManager.h
@@ -195,8 +195,8 @@ private:
 	void on(AdcCommand::STA, UserConnection*, const AdcCommand&) throw();
 
 	// TimerManagerListener
-	void on(TimerManagerListener::Second, uint32_t aTick) throw();	
-	void on(TimerManagerListener::Minute, uint32_t aTick) throw();	
+	void on(TimerManagerListener::Second, uint64_t aTick) throw();	
+	void on(TimerManagerListener::Minute, uint64_t aTick) throw();	
 
 };
